+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Extracted `piral-search` library
 * Added `validate` command for Piral instances (#69)
 * Added `validate` command for pilets (#70)
+* Support for extensions in Angular (#54), Hyperapp, Vue
 
 ## 0.7.0 (September 6, 2019)